Sends client user input to server.
This is the first step in getting keyboard
input translated to actions, followed by
these actions being sent to the server
over the transport layer.

It also has some messy changes that
start to track the frame id. The idea is
to tag frame ids on inputs to perform
reconcilliation of input due to latency.

This change still needs more work to
get the user input processed on the
server as well as lots of cleanup. I'm still
feeling out the ecsy interfaces and
trying to settle on some nice patterns.